fix(android): make font cap React 19-safe; dedupe native clamp
Address PR review feedback (#3216):

- React 19 ignores `defaultProps` on function components, so the previous
  Text/TextInput.defaultProps.maxFontSizeMultiplier assignments were a no-op
  and left iOS (and any JS-rendered text) with no app-wide font cap. Instead
  wrap each component's forwardRef render and inject maxFontSizeMultiplier
  into its props; explicit per-component values still take precedence.

- Extract the fontScale/densityDpi clamp into a shared DisplayScalingClamp
  helper called by both MainActivity and MainApplication, removing the
  duplicated logic and limits.
